% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECO.R
\name{DECO_LASSO_R}
\alias{DECO_LASSO_R}
\title{DECO Parallelized Algorithm}
\usage{
DECO_LASSO_R(Y, X, p, n, m, lambda, r, ncores = 1)
}
\arguments{
\item{Y}{gives the nx1 vector of observations we wish to approximate with a linear model of type Y = Xb + e}

\item{X}{gives the nxp matrix of regressors, each column corresponding to a different regressor}

\item{p}{is the column dimension of X [equivalently, p is the number of regressor variables]}

\item{n}{is the row dimension of X (and Y) [equivalently, n is the number of observations/individuals]}

\item{m}{is the number of groups/blocks you wish to split X into, denoted X(i) for 1 <= i <= m}

\item{lambda}{gives the (fixed) penalty magnitude in the LASSO fit of the algorithm}

\item{r}{is a tweaking parameter for making the inverse more robust (as we take inverse of XX + r*I)}

\item{ncores}{determines the number of cores used on each machine to parallelize computation}
}
\description{
This implementation uses only R functions. Refinement step is missing, lambda is fixed.
}
\details{
The algorithm is based on the description in "DECOrrelated feature space partitioning
         for distributed sparse regression" in Wang, Dunson, and Leng (2016) if lambda is fixed and
         LASSO is used as the penalized regression scheme. The rotated versions of Y and X the authors denote
         with Tilde are denoted as X* and Y* in the comments below
}
\note{
-This implementation is suboptimal in that X is already stored in the memory when we start the procedure.
      Ideally, one would give in only the LOCATION X is stored at and read it in chunkwise (thus allowing for
      larger matrices X, as was intended by the authors).
      -The notation #~PARALLEl~# will be introduced in the code whereever one may achieve signficiant gains from
      parallelizing
      -I could evaluate old expressions in the R version within the mcapply loops! ->saves memory as we write over old
      data
      -We cannot disturb variable order within the algorithm for output comparison reasons, thus reorder X columns before
      running DECO_LASSO (if important)
}
\author{
Samuel Davenport, Jack Carter, Giulio Morina, Jeremias Knoblauch
}

