% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intersect.R
\name{get_intersect}
\alias{get_intersect}
\title{Get the intersection of a list of sets}
\usage{
get_intersect(sets, type = c("row", "col", "vector"))
}
\arguments{
\item{sets}{Either a list of matrices with named rows and columns, or a list of named vectors}

\item{type}{The type of sets that you want to get the intersection of. Either "row", "col", or "vector"}
}
\value{
A vector of the intersection of the sets
}
\description{
Get the intersection of a list of sets
}
\examples{
s <- list(c("a","b","c","d"), c("f","e","d","c"))
get_intersect(s, type = "vector")

}
